+++ src/components/NavBar.js
@@ -9,9 +9,9 @@ const NavBar = () => {
       <ul className="navbar__list">
         <li><Link to="/" className="navbar__link">Home</Link></li>
         <li><Link to="/writing" className="navbar__link">Writing</Link></li>
-        <li><Link to="/meme" className="navbar__link">Meme's Works</Link></li>
-        {/* <li><Link to="/resume" className="navbar__link">Resume</Link></li>
         <li><Link to="/photography" className="navbar__link">Photography</Link></li>
+        {/* <li><Link to="/resume" className="navbar__link">Resume</Link></li>
+        <li><Link to="/meme" className="navbar__link">Meme's Works</Link></li>
         <li><Link to="/clothing" className="navbar__link">Clothing</Link></li>
         <li><Link to="/cooking" className="navbar__link">Cooking</Link></li> */}
       </ul>
