+++ src/components/Photography.css
@@ -8,37 +8,32 @@ h1 {
   margin-bottom: 20px;
 }
 
+/* Artist's Statement */
+.artist-statement {
+  max-width: 800px;
+  margin: 0 auto 40px auto;
+  font-size: 1.2em;
+  line-height: 1.6;
+  text-align: center;
+  color: #555;
+}
+
+/* Scrolling Gallery */
 .scrolling-gallery {
   margin-bottom: 40px;
 }
 
 .slide img {
   width: 100%;
-  height: auto; /* Ensure images maintain their aspect ratio */
-  max-height: 500px; /* Adjust this value to control the height */
+  height: auto;
+  max-height: calc(100vh - 200px); /* Nearly full-screen height */
 }
 
-.slick-slide {
-  padding: 0 10px; /* Add some padding between slides */
+.slick-slider {
+  width: 90%; /* Adjust gallery width */
+  margin: 0 auto;
 }
 
 .slick-dots {
-  bottom: 10px; /* Position dots at the bottom of the gallery */
-}
-
-.slick-prev, .slick-next {
-  z-index: 1; /* Ensure navigation arrows are on top of the gallery */
-}
-
-/* Redbubble Section */
-.redbubble-section {
-  margin-top: 50px;
-}
-
-.embed-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 500px; /* Adjust the height as needed */
+  bottom: 15px;
 }
