+++ src/components/Shops.css
@@ -1,0 +1,44 @@
+.shops {
+    padding: 20px;
+    font-family: Arial, sans-serif;
+  }
+  
+  h1 {
+    text-align: center;
+    margin-bottom: 20px;
+  }
+  
+  .scrolling-gallery {
+    margin-bottom: 40px;
+  }
+  
+  .slide img {
+    width: 100%;
+    height: auto; /* Ensure images maintain their aspect ratio */
+    max-height: 500px; /* Adjust this value to control the height */
+  }
+  
+  .slick-slide {
+    padding: 0 10px; /* Add some padding between slides */
+  }
+  
+  .slick-dots {
+    bottom: 10px; /* Position dots at the bottom of the gallery */
+  }
+  
+  .slick-prev, .slick-next {
+    z-index: 1; /* Ensure navigation arrows are on top of the gallery */
+  }
+  
+  /* Redbubble Section */
+  .redbubble-section {
+    margin-top: 50px;
+  }
+  
+  .embed-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height: 500px; /* Adjust the height as needed */
+  }
